refactor route-controller-manager to use ControlPlaneComponent abstraction
