--- control-plane-operator/controllers/hostedcontrolplane/manifests/openshift_route_cm.go
@@ -1,35 +1,0 @@
-package manifests
-
-import (
-	prometheusoperatorv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
-	appsv1 "k8s.io/api/apps/v1"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-)
-
-func OpenShiftRouteControllerManagerDeployment(ns string) *appsv1.Deployment {
-	return &appsv1.Deployment{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "openshift-route-controller-manager",
-			Namespace: ns,
-		},
-	}
-}
-
-func OpenShiftRouteControllerManagerService(ns string) *corev1.Service {
-	return &corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "openshift-route-controller-manager",
-			Namespace: ns,
-		},
-	}
-}
-
-func OpenShiftRouteControllerManagerServiceMonitor(ns string) *prometheusoperatorv1.ServiceMonitor {
-	return &prometheusoperatorv1.ServiceMonitor{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "openshift-route-controller-manager",
-			Namespace: ns,
-		},
-	}
-}
